Separate examples to own page RSC-568
