+++ gallery/src/data-visualization/examples/ReactVisDonutChartExample.tsx
@@ -4,22 +4,21 @@
  * the terms of the Eclipse Public License 2.0 which accompanies this
  * distribution and is available at https://www.eclipse.org/legal/epl-2.0/.
  */
-/* eslint-disable no-console */
-import { NxButton, NxLoadingSpinner } from '@sonatype/react-shared-components';
 import React, { useEffect, useState } from 'react';
+import { NxButton, NxLoadingSpinner, selectableColors } from '@sonatype/react-shared-components';
 import { Hint, RadialChart, RadialChartPoint } from 'react-vis';
-import { randomNumberGenerator } from '../util/jsUtil';
-import './ReactVis.scss';
+import { randomNumberGenerator } from '../../util/jsUtil';
+import '../ReactVis.scss';
 
 export default function ReactVisDonutChartExample() {
   const data: RadialChartPoint[] = [
-    {angle: 1, color: 'red', label: 'Severe', subLabel: '1'},
-    {angle: 5, color: 'yellow', label: 'Moderate', subLabel: '5'},
-    {angle: 2, color: 'green', label: 'Low', subLabel: '2'}
+    {angle: 1, color: selectableColors[3], label: 'Severe', subLabel: '1'},
+    {angle: 5, color: selectableColors[6], label: 'Moderate', subLabel: '5'},
+    {angle: 2, color: selectableColors[4], label: 'Low', subLabel: '2'}
   ];
 
-  const [chartData, setChartData] = useState<RadialChartPoint[] | null>(null);
   const [isLoading, setIsLoading] = useState(true);
+  const [chartData, setChartData] = useState<RadialChartPoint[] | null>(null);
   const [hintValue, setHintValue] = useState<RadialChartPoint | null>(null);
 
   // Simulate an async data load task
@@ -27,18 +26,18 @@ export default function ReactVisDonutChartExample() {
     setTimeout(() => {
       setChartData(data);
       setIsLoading(false);
-    }, 3000);
+    }, 1000);
   }, []);
 
-  //Creates random datapoints
+  // Simulate new data load on button click
   const updateData = () => {
     const tempData: RadialChartPoint[] = [];
     let randomNum = randomNumberGenerator(1, 10, 0);
-    tempData.push({angle: randomNum, color: 'red', label: 'Severe', subLabel: String(randomNum)});
+    tempData.push({angle: randomNum, color: selectableColors[3], label: 'Severe', subLabel: String(randomNum)});
     randomNum = randomNumberGenerator(1, 10, 0);
-    tempData.push({angle: randomNum, color: 'green', label: 'Low', subLabel: String(randomNum)});
+    tempData.push({angle: randomNum, color: selectableColors[4], label: 'Low', subLabel: String(randomNum)});
     randomNum = randomNumberGenerator(1, 10, 0);
-    tempData.push({angle: randomNum, color: 'yellow', label: 'Moderate', subLabel: String(randomNum)});
+    tempData.push({angle: randomNum, color: selectableColors[6], label: 'Moderate', subLabel: String(randomNum)});
     setChartData(tempData);
     setHintValue(null);
   };
